+++ Assets/Editor/Editoraaaaa.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a5ea7f54c8a82f44285f09b0cef07479
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
