+++ Assets/Scripts/Enemy/EnemyTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a74752561f061ef4b8eea50cecbbfa91
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
