+++ Assets/Scripts/Enemy/Pattarn.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9e4df6a8147da3d428f9ea8d88becd71
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
